Moved tests into the correct project, removed errant references, and fixed bugs exposed by actually running the tests
