+++ src/UniversalDiveDataFormat.Tests/Models/BuiltTests.cs
@@ -1,15 +1,15 @@
 using System.Xml.Serialization;
 using Shouldly;
 using UniversalDiveDataFormat.ExtensionMethods;
-using Xunit;
+using UniversalDiveDataFormat.Models;
 
-namespace UniversalDiveDataFormat.Models;
+namespace UniversalDiveDataFormat.Tests.Models;
 
 public class BuiltTests
 {
 	private const string Xml = """
 	                           <built>
-	                               <shipyard>Blohm & Voss</shipyard>
+	                               <shipyard>Blohm &amp; Voss</shipyard>
 	                               <launchingdate>
 	                                   <datetime>1943-06-14</datetime>
 	                               </launchingdate>
