+++ src/UniversalDiveDataFormat/Services/UddfDeserializer.cs
@@ -1,0 +1,6 @@
+namespace UniversalDiveDataFormat.Services;
+
+public class UddfDeserializer
+{
+	
+}
