+++ src/UniversalDiveDataFormat/UniversalDiveDataFormat.csproj
@@ -18,12 +18,6 @@
 
   <ItemGroup>
     <None Include="..\..\README.md" Pack="true" PackagePath="\"/>
-    <Reference Include="Shouldly">
-      <HintPath>..\..\..\..\.nuget\packages\shouldly\4.2.1\lib\net5.0\Shouldly.dll</HintPath>
-    </Reference>
-    <Reference Include="xunit.core">
-      <HintPath>..\..\..\..\.nuget\packages\xunit.extensibility.core\2.4.2\lib\netstandard1.1\xunit.core.dll</HintPath>
-    </Reference>
   </ItemGroup>
 
 </Project>
